drivers/flash: Add erase_requirement flag to flash_parameters.
The commit adds erase_requirement flag, to flash parameters, that
allows to figure out whether device requires erase prior to write
or not.

There are following Kconfig options added:
 FLASH_DEVICE_HAS_ERASE_REQUIREMENTS
 FLASH_DEVICE_HAS_NO_ERASE_REQUIREMENTS
that should be selected by device driver to indicate whether
devices served by driver needs erase prior to write.
The above Kconfigs are used to figure out whether app is built
for hardware that requires erase prior to write.
They can be also used to detect that it is attempted to build
some subsystem that will not work with provided hardware, for example
file system that has not been prepared to work with devices that
do not require erase.

Kconfig are also used with FLASH_EVAL_ERASE_REQUIREMENT(condition)
macro that allows to optimize out code that would never be executed
in a specific hardware configuration: for example branch serving
devices that do not require erase could be left out if no such
devices are supported by such system.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>