applications/asset_tracker_v2: Switch from native_posix to native_sim
This application did not work properly on native_posix,
as it was selecting the POSIX_CLOCK API (which would
call into the Linux APIs intead of Zephyr's).
Recently that configutation combination was forbidden
in native_posix, which makes this application build
fail for that target.

Let's switch from using native_posix as a test platform
to native_sim which is its replacement.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>